+++ src/plugins/replication-graphql/graphql-schema-from-rx-schema.ts
@@ -32,7 +32,9 @@ export type GraphQLSchemaFromRxSchemaInputSingleCollection = {
      * These fields of the document data
      * will be used for the checkpoint.
      */
-    checkpointFields: string[];
+    checkpointFields: {
+        [fieldName: string]: GraphQLParamType;
+    };
     ignoreInputKeys?: string[];
     ignoreOutputKeys?: string[];
     withRevisions?: boolean;