+++ test/unit.test.ts
@@ -23,6 +23,8 @@ import './unit/rx-storage-pouchdb.test';
 import './unit/rx-storage-lokijs.test';
 import './unit/rx-storage-dexie.test';
 
+// import './unit/server-graphql.test';
+import './unit/replication-graphql.test';
 
 
 import './unit/instance-of-check.test';
@@ -54,7 +56,6 @@ import './unit/hooks.test';
 import './unit/orm.test';
 import './unit/replication-protocol.test';
 import './unit/replication.test';
-import './unit/replication-graphql.test';
 import './unit/replication-couchdb.test';
 import './unit/population.test';
 import './unit/leader-election.test';